Cleanup P_SNIActionPerformer
Move function definitions into the .cc file so the hedaer file doesn't bring unrelated dependencies.
