+++ src/iocore/net/SSLSNIConfig.cc
@@ -30,7 +30,10 @@
  ****************************************************************************/
 
 #include "iocore/net/SSLSNIConfig.h"
-#include "P_SNIActionPerformer.h"
+#include "iocore/net/SNIActionItem.h"
+
+#include "P_SSLUtils.h"
+#include "P_SSLConfig.h"
 
 #include "tscore/Diags.h"
 #include "tscore/SimpleTokenizer.h"
